Java: GptUi: store temperature as Integer

+++ Java+/App+/GptUi/src/gptui/gpt/gcp/GcpApi.java
@@ -3,5 +3,5 @@ package gptui.gpt.gcp;
 import java.math.BigDecimal;
 
 public interface GcpApi {
-    String send(String content, BigDecimal temperature);
+    String send(String content, Integer temperature);
 }
